fix(web): debounce plan-affecting migration inputs to stop form reloads
Short name and template suffix genuinely re-run preflight, so debounce them and
keep the previous preview rendered during the background refetch instead of
reloading the form on every keystroke. Extract useDebouncedValue into a shared
hook. Display-only name/term edits already avoid refetching.
